requirements.txt
Bottleneck==1.3.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
numexpr==2.8.7
numpy==1.26.0
pandas==2.1.1
pip==23.3
pyaml==20.4.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==68.0.0
six==1.16.0
tzdata==2023.3
wheel==0.41.2
